backend/requirements.txt
python-barcode==0.14
Pillow>=10.0
django-cleanup==6.2.0
openpyxl>=3.0.0
